+++ src/gurobi_logtools/parsers/presolve.py
@@ -17,16 +17,13 @@ class PresolveParser(Parser):
     )
 
     # Possible intermediate patterns to be parsed
-    presolve_intermediate_patterns = [
+    model_stat_patterns = [
         re.compile(r"Model fingerprint: (?P<Fingerprint>.*)$"),
-        re.compile(
-            r"Variable types: (?P<PresolvedNumConVars>\d+) continuous, (?P<PresolvedNumIntVars>\d+) integer \((?P<PresolvedNumBinVars>\d+) binary\)$",
-        ),
-        re.compile(
+        re.compile(  # dev log
             r"Variable types: (?P<PresolvedNumBinVars>\d+) bin/(?P<PresolvedNumIntVars>\d+) gen[^/]*/(?P<PresolvedNumConVars>\d+) continuous",
         ),
         re.compile(
-            r"Semi-Variable types: (?P<PresolvedNumSemiContVars>\d+) continuous, (?P<PresolvedNumSemiIntVars>\d+) integer$",
+            r"Semi-Variable types: (?P<NumSemiContVars>\d+) continuous, (?P<NumSemiIntVars>\d+) integer$",
         ),
         re.compile(
             r"\s*QMatrix range\s*\[(?P<MinQCCoeff>[^,]+),\s*(?P<MaxQCCoeff>[^\]]+)\]",
@@ -60,6 +57,9 @@ class PresolveParser(Parser):
         re.compile(
             r"Concurrent MIP optimizer: (?P<ConcurrentJobs>\d+) concurrent instances \(\d+ threads per instance\)",
         ),
+    ]
+
+    presolve_intermediate_patterns = [
         re.compile(
             r"Presolved: (?P<PresolvedNumConstrs>\d+) (R|r)ows, (?P<PresolvedNumVars>\d+) (C|c)olumns, (?P<PresolvedNumNZs>\d+) (N|n)on(Z|z)ero(e?)s",
         ),
@@ -69,6 +69,11 @@ class PresolveParser(Parser):
         ),
         re.compile(r"Presolve time: (?P<PresolveTime>[\d\.]+)s"),
     ]
+
+    # Special case
+    variable_types = re.compile(
+        r"Variable types: (?P<NumConVars>\d+) continuous, (?P<NumIntVars>\d+) integer \((?P<NumBinVars>\d+) binary\)$",
+    )
 
     # Special case: model solved by presolve
     presolve_all_removed = re.compile(r"Presolve: All rows and columns removed")
@@ -82,6 +87,8 @@ class PresolveParser(Parser):
         """
         self._summary: Dict[str, Any] = {}
         self._started = False
+        self._post_presolve = False
+        self._original_variable_types_matched = False
         self._pretree_solution_parser = pretree_solution_parser
 
     def parse(self, line: str) -> ParseResult:
@@ -108,12 +115,33 @@ class PresolveParser(Parser):
         if parse_result := self._pretree_solution_parser.parse(line):
             return parse_result
 
-        for pattern in PresolveParser.presolve_intermediate_patterns:
+        for pattern in PresolveParser.model_stat_patterns:
             match = pattern.match(line)
             if match:
                 parse_result = typeconvert_groupdict(match)
                 self._summary.update(parse_result)
                 return ParseResult(parse_result)
+
+        for pattern in PresolveParser.presolve_intermediate_patterns:
+            match = pattern.match(line)
+            if match:
+                self._post_presolve = True
+                parse_result = typeconvert_groupdict(match)
+                self._summary.update(parse_result)
+                return ParseResult(parse_result)
+
+        if match := self.variable_types.match(line):
+            parse_result = typeconvert_groupdict(match)
+            if self._post_presolve:
+                parse_result = {
+                    "PresolvedNumConVars": parse_result["NumConVars"],
+                    "PresolvedNumIntVars": parse_result["NumIntVars"],
+                    "PresolvedNumBinVars": parse_result["NumBinVars"],
+                }
+            else:
+                self._original_variable_types_matched = True
+            self._summary.update(parse_result)
+            return ParseResult(parse_result)
 
         match = PresolveParser.presolve_all_removed.match(line)
         if match:
@@ -135,10 +163,10 @@ class PresolveParser(Parser):
             discrete_vars=sum(
                 summary.get(k, 0)
                 for k in (
-                    "PresolvedNumBinVars",
-                    "PresolvedNumIntVars",
-                    "PresolvedNumSemiContVars",
-                    "PresolvedNumSemiIntVars",
+                    "NumBinVars",
+                    "NumIntVars",
+                    "NumSemiContVars",
+                    "NumSemiIntVars",
                 )
             ),
             quad_nonzeros=summary.get("NumQNZs", 0),